refactor: delete test joke command
- delete generate joke method in gpt service
- move pdf tools class into helper folder
