+++ src/DjinniAIReplyBot.Application/DjinniAIReplyBot.Application.csproj
@@ -18,6 +18,7 @@
 
     <ItemGroup>
       <Folder Include="Models\ChatGpt\" />
+      <Folder Include="Models\Telegram\" />
     </ItemGroup>
 
 </Project>
